refactor(credits): streamline credit retrieval methods and remove deprecated service files
