+++ web-app/src/app/(app)/agents/[agentId]/jobs/layout.tsx
@@ -9,7 +9,7 @@ import {
 import DefaultLoading from "@/components/default-loading";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
 import { getAgentDescription, getAgentLegal, getAgentName } from "@/lib/db";
-import { AgentService, getAgentCreditsPrice, JobService } from "@/lib/services";
+import { AgentService, JobService } from "@/lib/services";
 
 import Footer from "./components/footer";
 import Header, { HeaderSkeleton } from "./components/header";
@@ -59,7 +59,7 @@ async function JobLayoutInner({ right, params, children }: JobLayoutProps) {
     return notFound();
   }
 
-  const agentCreditsPrice = await getAgentCreditsPrice(agent);
+  const agentCreditsPrice = await agentService.getAgentCreditsPrice(agent);
   const favoriteAgents = await agentService.getFavoriteAgents();
   const jobs =
     await JobService.getInstance().getJobsWithLimitedInformationByAgentId(
