--- web-app/src/lib/services/credit/index.ts
@@ -1,1 +1,0 @@
-export * from "./service";
