+++ web-app/src/lib/services/index.ts
@@ -1,7 +1,6 @@
 export { AgentService } from "./agent.service";
 export { AgentListService } from "./agentList.service";
 export { BaseService } from "./base.service";
-export * from "./credit";
 export { CreditCostService } from "./creditCost.service";
 export { CreditTransactionService } from "./creditTransaction.service";
 export { FiatTransactionService } from "./fiatTransaction.service";
